requirements.txt
pdfplumber==0.5.28
python-docx==0.8.11
beautifulsoup4==4.9.3
